Voucher statistics

+++ SIMS-Initial-Project-main/InitialProject/InitialProject/Dto/TourStatisticsDto.cs
@@ -7,8 +7,8 @@ public class TourStatisticsDto
     public int YouthCount{ get; set; }
     public int MiddleAgedCount { get; set; }
     public int OldPeopleCount{ get; set; }
-    public double WithVouchers { get; set; }
-    public double WithoutVouchers{ get; set; }
+    public int WithVouchers { get; set; }
+    public int WithoutVouchers{ get; set; }
 
 
     public TourStatisticsDto()
